# api/openapi.json
{
  "components": {
    "schemas": {
...
      "DesignChange": {
        "properties": {
          "revision": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "revision"
        ],
        "type": "object"
      },
      "DesignEdit": {
        "properties": {
          "command_id": {
            "type": "string"
          },
          "expected_revision": {
            "format": "int64",
            "type": "integer"
          },
          "ops": {
            "items": {
              "$ref": "#/components/schemas/DesignOp"
            },
            "type": "array"
          }
        },
        "required": [
          "command_id",
          "expected_revision",
          "ops"
        ],
        "type": "object"
      },
      "DesignOp": {
        "oneOf": [
          {
            "properties": {
              "before": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/OutcomeId"
                  }
                ]
              },
              "content": {
                "$ref": "#/components/schemas/OutcomeContent"
              },
              "id": {
                "$ref": "#/components/schemas/OutcomeId"
              },
              "kind": {
                "enum": [
                  "OutcomeCreate"
                ],
                "type": "string"
              },
              "parent": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/OutcomeId"
                  }
                ]
              }
            },
            "required": [
              "id",
              "content",
              "kind"
            ],
            "type": "object"
          },
          {
            "properties": {
              "content": {
                "$ref": "#/components/schemas/OutcomeContent"
              },
              "id": {
                "$ref": "#/components/schemas/OutcomeId"
              },
              "kind": {
                "enum": [
                  "OutcomePut"
                ],
                "type": "string"
              }
            },
            "required": [
              "id",
              "content",
              "kind"
            ],
            "type": "object"
          },
          {
            "properties": {
              "before": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/OutcomeId"
                  }
                ]
              },
              "id": {
                "$ref": "#/components/schemas/OutcomeId"
              },
              "kind": {
                "enum": [
                  "OutcomeMove"
                ],
                "type": "string"
              },
              "parent": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/OutcomeId"
                  }
                ]
              }
            },
            "required": [
              "id",
              "kind"
            ],
            "type": "object"
          },
          {
            "properties": {
              "kind": {
                "enum": [
                  "OutcomePartPut"
                ],
                "type": "string"
              },
              "outcome": {
                "$ref": "#/components/schemas/OutcomeId"
              },
              "part": {
                "$ref": "#/components/schemas/PartId"
              }
            },
            "required": [
              "outcome",
              "part",
              "kind"
            ],
            "type": "object"
          },
          {
            "properties": {
              "kind": {
                "enum": [
                  "OutcomePartRemove"
                ],
                "type": "string"
              },
              "outcome": {
                "$ref": "#/components/schemas/OutcomeId"
              },
              "part": {
                "$ref": "#/components/schemas/PartId"
              }
            },
            "required": [
              "outcome",
              "part",
              "kind"
            ],
            "type": "object"
          },
          {
            "properties": {
              "content": {
                "$ref": "#/components/schemas/VisionContent"
              },
              "kind": {
                "enum": [
                  "VisionPut"
                ],
                "type": "string"
              }
            },
            "required": [
              "content",
              "kind"
            ],
            "type": "object"
          },
          {
            "properties": {
              "before": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/PartId"
                  }
                ]
              },
              "content": {
                "$ref": "#/components/schemas/PartContent"
              },
              "id": {
                "$ref": "#/components/schemas/PartId"
              },
              "kind": {
                "enum": [
                  "PartCreate"
                ],
                "type": "string"
              },
              "parent": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/PartId"
                  }
                ]
              }
            },
            "required": [
              "id",
              "content",
              "kind"
            ],
            "type": "object"
          },
          {
            "properties": {
              "content": {
                "$ref": "#/components/schemas/PartContent"
              },
              "id": {
                "$ref": "#/components/schemas/PartId"
              },
              "kind": {
                "enum": [
                  "PartPut"
                ],
                "type": "string"
              }
            },
            "required": [
              "id",
              "content",
              "kind"
            ],
            "type": "object"
          },
          {
            "properties": {
              "before": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/PartId"
                  }
                ]
              },
              "id": {
                "$ref": "#/components/schemas/PartId"
              },
              "kind": {
                "enum": [
                  "PartMove"
                ],
                "type": "string"
              },
              "parent": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/PartId"
                  }
                ]
              }
            },
            "required": [
              "id",
              "kind"
            ],
            "type": "object"
          },
          {
            "properties": {
              "anchor": {
                "$ref": "#/components/schemas/DesignAnchor"
              },
              "kind": {
                "enum": [
                  "PlanLinkPut"
                ],
                "type": "string"
              },
              "plan": {
                "$ref": "#/components/schemas/PlanId"
              }
            },
            "required": [
              "anchor",
              "plan",
              "kind"
            ],
            "type": "object"
          },
          {
            "properties": {
              "anchor": {
                "$ref": "#/components/schemas/DesignAnchor"
              },
              "kind": {
                "enum": [
                  "PlanLinkRemove"
                ],
                "type": "string"
              },
              "plan": {
                "$ref": "#/components/schemas/PlanId"
              }
            },
            "required": [
              "anchor",
              "plan",
              "kind"
            ],
            "type": "object"
          }
        ]
      },
...
      "ErrorBody": {
        "description": "The body of every error this API answers: the stable code a client\nmatches on (spec §3.4), and a human message nobody should match on. A\nrequest axum rejects before a handler runs (malformed JSON, a missing query\nparameter) gets one too, through\n[`rejections_as_error_bodies`].",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "current_revision": {
            "description": "`REVISION_CONFLICT` only: the resource's revision now, to read again at.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "problems": {
            "description": "`WORKFLOW_VALIDATION_FAILED` only: each problem, the same sentences\n`message` joins.",
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "InstructionsVersion": {
        "description": "One saved version of a project's instructions. `id` is what an\n`agent_invocation` records (§13.15); a client sees the number instead.",
        "properties": {
          "body": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "number": {
            "description": "1, 2, 3 … within the project.",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "number",
          "body",
          "created_at"
        ],
        "type": "object"
      },
...
      "OutcomeId": {
        "format": "uuid",
        "type": "string"
      },
      "OutcomePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Outcome"
            },
            "type": "array"
          },
          "next": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OutcomeId"
              }
            ]
          },
          "revision": {
            "$ref": "#/components/schemas/i64"
          }
        },
        "required": [
          "revision",
          "items"
        ],
        "type": "object"
      },
      "OutcomeView": {
        "properties": {
          "ancestors": {
            "items": {
              "$ref": "#/components/schemas/Outcome"
            },
            "type": "array"
          },
          "outcome": {
            "$ref": "#/components/schemas/Outcome"
          },
          "parts": {
            "items": {
              "$ref": "#/components/schemas/PartId"
            },
            "type": "array"
          },
          "plans": {
            "items": {
              "$ref": "#/components/schemas/PlanId"
            },
            "type": "array"
          },
          "revision": {
            "$ref": "#/components/schemas/i64"
          }
        },
        "required": [
          "revision",
          "outcome",
          "ancestors",
          "parts",
          "plans"
        ],
        "type": "object"
      },
...
      "PartId": {
        "format": "uuid",
        "type": "string"
      },
      "PartPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Part"
            },
            "type": "array"
          },
          "next": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PartId"
              }
            ]
          },
          "revision": {
            "$ref": "#/components/schemas/i64"
          }
        },
        "required": [
          "revision",
          "items"
        ],
        "type": "object"
      },
      "PartView": {
        "properties": {
          "ancestors": {
            "items": {
              "$ref": "#/components/schemas/Part"
            },
            "type": "array"
          },
          "part": {
            "$ref": "#/components/schemas/Part"
          },
          "plans": {
            "items": {
              "$ref": "#/components/schemas/PlanId"
            },
            "type": "array"
          },
          "revision": {
            "$ref": "#/components/schemas/i64"
          }
        },
        "required": [
          "revision",
          "part",
          "ancestors",
          "plans"
        ],
        "type": "object"
      },
...
      "PlanId": {
        "format": "uuid",
        "type": "string"
      },
...
      "ProjectId": {
        "format": "uuid",
        "type": "string"
      },
...
      "SaveInstructions": {
        "properties": {
          "body": {
            "description": "The instructions, as the person wrote them.",
            "type": "string"
          },
          "command_id": {
            "description": "The idempotency key (spec §13.5), scoped to the project.",
            "type": "string"
          }
        },
        "required": [
          "command_id",
          "body"
        ],
        "type": "object"
      },
...
      "VisionView": {
        "properties": {
          "content": {
            "$ref": "#/components/schemas/VisionContent"
          },
          "revision": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "content"
        ],
        "type": "object"
      },
...
      "i64": {
        "format": "int64",
        "type": "integer"
      }
    }
  },
  "info": {
    "description": "The local daemon's HTTP API. Every client is on another origin; only origins passed with `--allow-origin` may call it. There is no authentication: the daemon binds to loopback (spec §1, OPEN block on remote access). Errors answered by the API carry `ErrorBody`, whose `code` is stable; a request axum rejects before a handler runs (malformed JSON, a missing parameter) is answered in plain text.",
    "license": {
      "name": ""
    },
    "title": "shadows",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/{id}/design/edits": {
      "post": {
        "operationId": "edit",
        "parameters": [
          {
            "description": "The project",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DesignEdit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesignChange"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "REVISION_CONFLICT or COMMAND_CONFLICT"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "design"
        ]
      }
    },
    "/api/projects/{id}/design/outcomes": {
      "get": {
        "operationId": "outcomes",
        "parameters": [
          {
            "description": "The project",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectId"
            }
          },
          {
            "in": "query",
            "name": "parent",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OutcomeId"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OutcomeId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutcomePage"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "design"
        ]
      }
    },
    "/api/projects/{id}/design/outcomes/{outcome}": {
      "get": {
        "operationId": "outcome",
        "parameters": [
          {
            "description": "The project",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectId"
            }
          },
          {
            "description": "The outcome",
            "in": "path",
            "name": "outcome",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OutcomeId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutcomeView"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "design"
        ]
      }
    },
    "/api/projects/{id}/design/parts": {
      "get": {
        "operationId": "parts",
        "parameters": [
          {
            "description": "The project",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectId"
            }
          },
          {
            "in": "query",
            "name": "parent",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PartId"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PartId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartPage"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "design"
        ]
      }
    },
    "/api/projects/{id}/design/parts/{part}": {
      "get": {
        "operationId": "part",
        "parameters": [
          {
            "description": "The project",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectId"
            }
          },
          {
            "description": "The part",
            "in": "path",
            "name": "part",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PartId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartView"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "design"
        ]
      }
    },
    "/api/projects/{id}/design/vision": {
      "get": {
        "operationId": "vision",
        "parameters": [
          {
            "description": "The project",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisionView"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "design"
        ]
      }
    },
...
    "/api/projects/{id}/planner-instructions": {
      "get": {
        "operationId": "get_instructions",
        "parameters": [
          {
            "description": "The project",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/InstructionsVersion"
                    }
                  ]
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "STORAGE_UNAVAILABLE"
          }
        },
        "summary": "The project's current instructions — its highest-numbered version — or\n`null` before the first save.",
        "tags": [
          "projects"
        ]
      },
      "put": {
        "operationId": "save_instructions",
        "parameters": [
          {
            "description": "The project",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveInstructions"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstructionsVersion"
                }
              }
            },
            "description": "Saved, or the replay of the same command"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "INVALID_COMMAND: no such project"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "COMMAND_CONFLICT"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "STORAGE_UNAVAILABLE"
          }
        },
        "summary": "Saves the project's instructions as its next version; nothing earlier is\noverwritten. A running Planner is told of the change before its next turn\n(§13.8), never mid-turn.",
        "tags": [
          "projects"
        ]
      }
    },
...
  },
...
}